refactor: move popup src to components
