+++ src/routes/+error.svelte
@@ -1,6 +1,7 @@
 <script>
     import SmackMagicWrapper from "$lib/ui/layouts/shop/SmackMagicWrapper.svelte";
 </script>
+
 <SmackMagicWrapper>
     <p>You shouldn't be here.</p>
     <a href="/">Back</a>
